Additional bugfixes for floating point number of NOPs.  Typechecking did not error on these.  Removed defaults from templates since mis-casting a void pointer was part of the problem.  Fixed some formatting.

+++ core-core/closed_loop_helpers.h
@@ -82,8 +82,8 @@ public:
     int64_t controlChecks;
     int64_t speed_up_count;
     int64_t slow_down_count;
-    int64_t serverNops;
-    int64_t clientNops;
+    float serverNops;
+    float clientNops;
     bool errored;
     int64_t transaction;
